[connectors-sdk] rename and move stuff
