+++ connectors-sdk/connectors_sdk/__init__.py
@@ -2,14 +2,14 @@
 
 __version__ = "0.1.0"
 
+from connectors_sdk.connectors.external_import._work_manager import WorkManager
 from connectors_sdk.connectors.external_import.base_data_processor import (
     BaseDataProcessor,
 )
-from connectors_sdk.connectors.external_import.base_external_import_connector import (
-    BaseExternalImportConnector,
+from connectors_sdk.connectors.external_import.external_import_connector import (
+    ExternalImportConnector,
 )
 from connectors_sdk.connectors.external_import.logger import ConnectorLogger
-from connectors_sdk.connectors.external_import.work_manager import WorkManager
 from connectors_sdk.settings.annotated_types import (
     DatetimeFromIsoString,
     ListFromString,
@@ -55,7 +55,7 @@ __all__ = [
     # Connector States
     "ExternalImportConnectorState",
     # Connector base classes
-    "BaseExternalImportConnector",
+    "ExternalImportConnector",
     "ConnectorLogger",
     "BaseDataProcessor",
     "WorkManager",
