+++ connectors-sdk/connectors_sdk/connectors/external_import/__init__.py
@@ -4,5 +4,5 @@ This module provides the foundational classes for building OpenCTI connectors:
 - ConnectorLogger: Logging wrapper to avoid direct pycti dependency
 - WorkManager: Work lifecycle management (initiate, send bundles, complete)
 - BaseDataProcessor: Abstract base class for data collection/processing
-- BaseExternalImportConnector: Full orchestration for external import connectors
+- ExternalImportConnector: Full orchestration for external import connectors
 """
